moving the keycode map to a separate file
/ref https://github.com/laurent22/joplin/pull/4022#issuecomment-726263640
